+++ packages/create-cosmos-app/src/git-question-template.ts
@@ -5,12 +5,12 @@ const glob = require('glob').sync;
 const fs = require('fs');
 const path = require('path');
 
-export const createApp = (repo: string) => {
+export const createQuestionTemplate = (repo: string) => {
     return async argv => {
         const { name } = await prompt([
             {
                 name: 'name',
-                message: 'Enter your new module name',
+                message: 'Enter your new app name',
                 required: true,
             }
         ], argv);
@@ -129,9 +129,9 @@ export const createApp = (repo: string) => {
         shell.rm('-rf', '.questions.json');
 
         console.log(`
-
-        |||
-       (o o)
+        
+           |||
+          (o o)
       ooO--(_)--Ooo-
 
       ✨ Great work!