+++ packages/create-cosmos-app/src/templates/typescript-npm-module.ts
@@ -1,3 +1,0 @@
-import { createApp } from "../git-question-template";
-
-export default createApp('git@github.com:supercosmonauts/cca-ts-module.git');