+++ packages/create-cosmos-app/CHANGELOG.md
@@ -3,6 +3,86 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.2](https://github.com/cosmology-tech/create-cosmos-app/compare/create-cosmos-app@0.2.1...create-cosmos-app@0.2.2) (2022-08-27)
+
+**Note:** Version bump only for package create-cosmos-app
+
+
+
+
+
+## [0.2.1](https://github.com/cosmology-tech/create-cosmos-app/compare/create-cosmos-app@0.2.0...create-cosmos-app@0.2.1) (2022-08-25)
+
+**Note:** Version bump only for package create-cosmos-app
+
+
+
+
+
+# [0.2.0](https://github.com/cosmology-tech/create-cosmos-app/compare/create-cosmos-app@0.1.0...create-cosmos-app@0.2.0) (2022-08-25)
+
+**Note:** Version bump only for package create-cosmos-app
+
+
+
+
+
+# [0.1.0](https://github.com/cosmology-tech/create-cosmos-app/compare/create-cosmos-app@0.0.12...create-cosmos-app@0.1.0) (2022-08-25)
+
+**Note:** Version bump only for package create-cosmos-app
+
+
+
+
+
+## [0.0.12](https://github.com/cosmology-tech/create-cosmos-app/compare/create-cosmos-app@0.0.11...create-cosmos-app@0.0.12) (2022-08-25)
+
+**Note:** Version bump only for package create-cosmos-app
+
+
+
+
+
+## [0.0.11](https://github.com/cosmology-tech/create-cosmos-app/compare/create-cosmos-app@0.0.10...create-cosmos-app@0.0.11) (2022-08-25)
+
+**Note:** Version bump only for package create-cosmos-app
+
+
+
+
+
+## [0.0.10](https://github.com/cosmology-tech/create-cosmos-app/compare/create-cosmos-app@0.0.9...create-cosmos-app@0.0.10) (2022-08-24)
+
+**Note:** Version bump only for package create-cosmos-app
+
+
+
+
+
+## [0.0.9](https://github.com/cosmology-tech/create-cosmos-app/compare/create-cosmos-app@0.0.8...create-cosmos-app@0.0.9) (2022-08-24)
+
+**Note:** Version bump only for package create-cosmos-app
+
+
+
+
+
+## [0.0.8](https://github.com/cosmology-tech/create-cosmos-app/compare/create-cosmos-app@0.0.7...create-cosmos-app@0.0.8) (2022-08-24)
+
+**Note:** Version bump only for package create-cosmos-app
+
+
+
+
+
+## [0.0.7](https://github.com/cosmology-tech/create-cosmos-app/compare/create-cosmos-app@0.0.6...create-cosmos-app@0.0.7) (2022-08-24)
+
+**Note:** Version bump only for package create-cosmos-app
+
+
+
+
+
 ## [0.0.6](https://github.com/cosmology-tech/create-cosmos-app/compare/create-cosmos-app@0.0.5...create-cosmos-app@0.0.6) (2022-08-18)
 
 **Note:** Version bump only for package create-cosmos-app